사이드 바 및 jsp들 높이 수정

+++ khis/src/main/java/com/kh/khis/interview_evaluation/model/service/EvaluationService.java
@@ -41,9 +41,9 @@ public interface EvaluationService {
 
 	int passnonpass(PassOrNonPass pass);
 
-	List<Interviewer> selectInterviewerList(int limit, int offset, int co_code);
+	List<Interviewer> selectInterviewerList(int limit, int offset, long co_code);
 
-	int selectInterviewerCount(int co_code);
+	int selectInterviewerCount(long co_code);
 
 	List<Assigned> selectInterviewerDetail(Map<String, Object> param);
